Make storage upgrade robust against changing struct layout.

+++ include/keepkey/board/keepkey_flash.h
@@ -28,7 +28,7 @@
 
 /* === Functions =========================================================== */
 
-uint32_t flash_write_helper(Allocation group);
+intptr_t flash_write_helper(Allocation group);
 void flash_erase(Allocation group);
 void flash_erase_word(Allocation group);
 bool flash_write(Allocation group, uint32_t offset, uint32_t len, uint8_t* data);
